// simulator/src/api/schemas/generator/types.ts
import {
	GraphQLObjectType,
	GraphQLID,
	GraphQLFloat,
	GraphQLBoolean,
} from 'graphql';

const typeName = 'Generator';

const GeneratorType = new GraphQLObjectType({
	name: typeName,
	description: `A ${typeName} object type.`,
	fields: {
		id: {
			type: GraphQLID,
			description: `The id of the ${typeName}.`
		},
		baseOutput: {
			type: GraphQLFloat,
			description: `The electricity output of the ${typeName} when working at 100% efficiency in watt seconds.`
		},
		isBroken: {
			type: GraphQLBoolean,
			description: `Whether or not the ${typeName} is broken.`
		},
		pollution: {
			type: GraphQLFloat,
			description: `Pollution units generated per watt seconds generated by the ${typeName}.`
		}
	}
});

export {
	typeName,
	GeneratorType
};